fix erase and remove, update internal names
vector_remove and vector_erase have been broken for a long time, but it doesn't seem like many people noticed, so I'm not too worried about breaking the api. The docs say that the parameter names determine whether or not to take the address of the vector pointer, so in that sense, the api is not changing.